Update in response to feedback
