--- addons/shot_counter/CfgDebriefingSections.hpp
@@ -1,12 +1,0 @@
-// Copy these to your mission's description.ext
-class CfgDebriefingSections {
-    class arc_cfg_shot_counter_countTeams {
-        title = "Munitions Expended by Teams";
-        variable = "arc_cfg_shot_counter_countTextTeams";
-    };
-
-    class arc_cfg_shot_counter_countPlayer {
-        title = "Munitions Expended by You";
-        variable = "arc_cfg_shot_counter_countTextPlayer";
-    };
-};
