+++ addons/shot_counter/config.cpp
@@ -12,5 +12,4 @@ class CfgPatches {
     };
 };
 
-#include "CfgDebriefingSections.hpp"
 #include "CfgEventHandlers.hpp"
